// src/generated/ast.ts
export type ClassifierConstraint = DefExpression | InvariantConstraint;

export const ClassifierConstraint = 'ClassifierConstraint';

export function isClassifierConstraint(item: unknown): item is ClassifierConstraint {
    return reflection.isInstance(item, ClassifierConstraint);
}

export type CoclContext = CoclDerived | CoclReferenceFilter | CoclValidation;

export const CoclContext = 'CoclContext';
// ...
export type ContextDeclaration = ClassifierContext | CoclContext | OperationContext | PropertyContext;

export const ContextDeclaration = 'ContextDeclaration';

export function isContextDeclaration(item: unknown): item is ContextDeclaration {
    return reflection.isInstance(item, ContextDeclaration);
}

export type Expression = ArrowExpression | BinaryExpression | BooleanLiteral | CollectionLiteral | DotExpression | IfExpression | IntegerLiteral | InvalidLiteral | LetExpression | NullLiteral | RealLiteral | SelfExpression | StringLiteral | TupleLiteral | TypeExpLiteral | UnaryExpression | VariableExpression;

export const Expression = 'Expression';

export function isExpression(item: unknown): item is Expression {
    return reflection.isInstance(item, Expression);
}

export type OperationCondition = BodyCondition | PostCondition | PreCondition;

export const OperationCondition = 'OperationCondition';

export function isOperationCondition(item: unknown): item is OperationCondition {
    return reflection.isInstance(item, OperationCondition);
}

export type PropertyConstraint = DeriveConstraint | InitConstraint;

export const PropertyConstraint = 'PropertyConstraint';

export function isPropertyConstraint(item: unknown): item is PropertyConstraint {
    return reflection.isInstance(item, PropertyConstraint);
}

export type QualifiedName = string;

export function isQualifiedName(item: unknown): item is QualifiedName {
    return typeof item === 'string';
}

export type TypeExpression = CollectionType | SimpleType | TupleType;

export const TypeExpression = 'TypeExpression';

export function isTypeExpression(item: unknown): item is TypeExpression {
    return reflection.isInstance(item, TypeExpression);
}

export interface ArrowExpression extends langium.AstNode {
    readonly $container: ArrowExpression | BinaryExpression | BodyCondition | CoclDerived | CoclReferenceFilter | CoclValidation | CollectionItem | DefExpression | DeriveConstraint | DotExpression | IfExpression | InitConstraint | InvariantConstraint | IteratorBody | LetExpression | PostCondition | PreCondition | TuplePart | UnaryExpression;
    readonly $type: 'ArrowExpression';
    arguments: Array<Expression | IteratorBody>;
    operation: string;
    source: Expression;
}

export const ArrowExpression = 'ArrowExpression';

export function isArrowExpression(item: unknown): item is ArrowExpression {
    return reflection.isInstance(item, ArrowExpression);
}

export interface BinaryExpression extends langium.AstNode {
    readonly $container: ArrowExpression | BinaryExpression | BodyCondition | CoclDerived | CoclReferenceFilter | CoclValidation | CollectionItem | DefExpression | DeriveConstraint | DotExpression | IfExpression | InitConstraint | InvariantConstraint | IteratorBody | LetExpression | PostCondition | PreCondition | TuplePart | UnaryExpression;
    readonly $type: 'BinaryExpression';
    left: Expression;
    op: '*' | '+' | '-' | '/' | '<' | '<=' | '<>' | '=' | '>' | '>=' | 'and' | 'div' | 'implies' | 'mod' | 'or' | 'xor';
    right: Expression;
}

export const BinaryExpression = 'BinaryExpression';

export function isBinaryExpression(item: unknown): item is BinaryExpression {
    return reflection.isInstance(item, BinaryExpression);
}

export interface BodyCondition extends langium.AstNode {
    readonly $container: OperationContext;
    readonly $type: 'BodyCondition';
    expression: Expression;
    name?: string;
}

export const BodyCondition = 'BodyCondition';

export function isBodyCondition(item: unknown): item is BodyCondition {
    return reflection.isInstance(item, BodyCondition);
}

export interface BooleanLiteral extends langium.AstNode {
    readonly $container: ArrowExpression | BinaryExpression | BodyCondition | CoclDerived | CoclReferenceFilter | CoclValidation | CollectionItem | DefExpression | DeriveConstraint | DotExpression | IfExpression | InitConstraint | InvariantConstraint | IteratorBody | LetExpression | PostCondition | PreCondition | TuplePart | UnaryExpression;
    readonly $type: 'BooleanLiteral';
    value: 'false' | 'true';
}

export const BooleanLiteral = 'BooleanLiteral';

export function isBooleanLiteral(item: unknown): item is BooleanLiteral {
    return reflection.isInstance(item, BooleanLiteral);
}

export interface ClassifierContext extends langium.AstNode {
    readonly $container: OclDocument;
    readonly $type: 'ClassifierContext';
    constraints: Array<ClassifierConstraint>;
    type: QualifiedName;
}

export const ClassifierContext = 'ClassifierContext';

export function isClassifierContext(item: unknown): item is ClassifierContext {
    return reflection.isInstance(item, ClassifierContext);
}

export interface CoclDerived extends langium.AstNode {
    readonly $container: OclDocument;
    readonly $type: 'CoclDerived';
    expression: Expression;
    path: QualifiedName;
}

export const CoclDerived = 'CoclDerived';

export function isCoclDerived(item: unknown): item is CoclDerived {
    return reflection.isInstance(item, CoclDerived);
}

export interface CoclReferenceFilter extends langium.AstNode {
    readonly $container: OclDocument;
    readonly $type: 'CoclReferenceFilter';
    expression: Expression;
    path: QualifiedName;
}

export const CoclReferenceFilter = 'CoclReferenceFilter';

export function isCoclReferenceFilter(item: unknown): item is CoclReferenceFilter {
    return reflection.isInstance(item, CoclReferenceFilter);
}

export interface CoclValidation extends langium.AstNode {
    readonly $container: OclDocument;
    readonly $type: 'CoclValidation';
    expression: Expression;
    path: QualifiedName;
}

export const CoclValidation = 'CoclValidation';

export function isCoclValidation(item: unknown): item is CoclValidation {
    return reflection.isInstance(item, CoclValidation);
}

export interface CollectionItem extends langium.AstNode {
    readonly $container: CollectionLiteral;
    readonly $type: 'CollectionItem';
    first: Expression;
    last?: Expression;
}

export const CollectionItem = 'CollectionItem';

export function isCollectionItem(item: unknown): item is CollectionItem {
    return reflection.isInstance(item, CollectionItem);
}

export interface CollectionLiteral extends langium.AstNode {
    readonly $container: ArrowExpression | BinaryExpression | BodyCondition | CoclDerived | CoclReferenceFilter | CoclValidation | CollectionItem | DefExpression | DeriveConstraint | DotExpression | IfExpression | InitConstraint | InvariantConstraint | IteratorBody | LetExpression | PostCondition | PreCondition | TuplePart | UnaryExpression;
    readonly $type: 'CollectionLiteral';
    items: Array<CollectionItem>;
    kind: CollectionKind;
}

export const CollectionLiteral = 'CollectionLiteral';

export function isCollectionLiteral(item: unknown): item is CollectionLiteral {
    return reflection.isInstance(item, CollectionLiteral);
}

export interface CollectionType extends langium.AstNode {
    readonly $container: CollectionType | DefExpression | IteratorBody | LetExpression | OperationContext | Parameter | PropertyContext | TupleTypePart;
    readonly $type: 'CollectionType';
    elementType: TypeExpression;
    kind: CollectionKind;
}

export const CollectionType = 'CollectionType';

export function isCollectionType(item: unknown): item is CollectionType {
    return reflection.isInstance(item, CollectionType);
}

export interface DefExpression extends langium.AstNode {
    readonly $container: ClassifierContext;
    readonly $type: 'DefExpression';
    body: Expression;
    name: string;
    parameters: Array<Parameter>;
    returnType?: TypeExpression;
    type?: TypeExpression;
}

export const DefExpression = 'DefExpression';

export function isDefExpression(item: unknown): item is DefExpression {
    return reflection.isInstance(item, DefExpression);
}

export interface DeriveConstraint extends langium.AstNode {
    readonly $container: PropertyContext;
    readonly $type: 'DeriveConstraint';
    expression: Expression;
}

export const DeriveConstraint = 'DeriveConstraint';

export function isDeriveConstraint(item: unknown): item is DeriveConstraint {
    return reflection.isInstance(item, DeriveConstraint);
}

export interface DotExpression extends langium.AstNode {
    readonly $container: ArrowExpression | BinaryExpression | BodyCondition | CoclDerived | CoclReferenceFilter | CoclValidation | CollectionItem | DefExpression | DeriveConstraint | DotExpression | IfExpression | InitConstraint | InvariantConstraint | IteratorBody | LetExpression | PostCondition | PreCondition | TuplePart | UnaryExpression;
    readonly $type: 'DotExpression';
    arguments: Array<Expression>;
    feature: string;
    source: Expression;
}

export const DotExpression = 'DotExpression';

export function isDotExpression(item: unknown): item is DotExpression {
    return reflection.isInstance(item, DotExpression);
}

export interface IfExpression extends langium.AstNode {
    readonly $container: ArrowExpression | BinaryExpression | BodyCondition | CoclDerived | CoclReferenceFilter | CoclValidation | CollectionItem | DefExpression | DeriveConstraint | DotExpression | IfExpression | InitConstraint | InvariantConstraint | IteratorBody | LetExpression | PostCondition | PreCondition | TuplePart | UnaryExpression;
    readonly $type: 'IfExpression';
    condition: Expression;
    elseExpression: Expression;
    thenExpression: Expression;
}

export const IfExpression = 'IfExpression';

export function isIfExpression(item: unknown): item is IfExpression {
    return reflection.isInstance(item, IfExpression);
}

export interface InitConstraint extends langium.AstNode {
    readonly $container: PropertyContext;
    readonly $type: 'InitConstraint';
    expression: Expression;
}

export const InitConstraint = 'InitConstraint';

export function isInitConstraint(item: unknown): item is InitConstraint {
    return reflection.isInstance(item, InitConstraint);
}

export interface IntegerLiteral extends langium.AstNode {
    readonly $container: ArrowExpression | BinaryExpression | BodyCondition | CoclDerived | CoclReferenceFilter | CoclValidation | CollectionItem | DefExpression | DeriveConstraint | DotExpression | IfExpression | InitConstraint | InvariantConstraint | IteratorBody | LetExpression | PostCondition | PreCondition | TuplePart | UnaryExpression;
    readonly $type: 'IntegerLiteral';
    value: number;
}

export const IntegerLiteral = 'IntegerLiteral';

export function isIntegerLiteral(item: unknown): item is IntegerLiteral {
    return reflection.isInstance(item, IntegerLiteral);
}

export interface InvalidLiteral extends langium.AstNode {
    readonly $container: ArrowExpression | BinaryExpression | BodyCondition | CoclDerived | CoclReferenceFilter | CoclValidation | CollectionItem | DefExpression | DeriveConstraint | DotExpression | IfExpression | InitConstraint | InvariantConstraint | IteratorBody | LetExpression | PostCondition | PreCondition | TuplePart | UnaryExpression;
    readonly $type: 'InvalidLiteral';
}

export const InvalidLiteral = 'InvalidLiteral';

export function isInvalidLiteral(item: unknown): item is InvalidLiteral {
    return reflection.isInstance(item, InvalidLiteral);
}

export interface InvariantConstraint extends langium.AstNode {
    readonly $container: ClassifierContext;
    readonly $type: 'InvariantConstraint';
    expression: Expression;
    name?: string;
}

export const InvariantConstraint = 'InvariantConstraint';

export function isInvariantConstraint(item: unknown): item is InvariantConstraint {
    return reflection.isInstance(item, InvariantConstraint);
}

export interface IteratorBody extends langium.AstNode {
    readonly $container: ArrowExpression;
    readonly $type: 'IteratorBody';
    body: Expression;
    iteratorName: string;
    iteratorName2?: string;
    iteratorType?: TypeExpression;
    iteratorType2?: TypeExpression;
}

export const IteratorBody = 'IteratorBody';

export function isIteratorBody(item: unknown): item is IteratorBody {
    return reflection.isInstance(item, IteratorBody);
}

export interface LetExpression extends langium.AstNode {
    readonly $container: ArrowExpression | BinaryExpression | BodyCondition | CoclDerived | CoclReferenceFilter | CoclValidation | CollectionItem | DefExpression | DeriveConstraint | DotExpression | IfExpression | InitConstraint | InvariantConstraint | IteratorBody | LetExpression | PostCondition | PreCondition | TuplePart | UnaryExpression;
    readonly $type: 'LetExpression';
    body: Expression;
    name: string;
    type?: TypeExpression;
    value: Expression;
}

export const LetExpression = 'LetExpression';
// ...
export interface OclDocument extends langium.AstNode {
    readonly $type: 'OclDocument';
    contexts: Array<ContextDeclaration>;
}

export const OclDocument = 'OclDocument';

export function isOclDocument(item: unknown): item is OclDocument {
    return reflection.isInstance(item, OclDocument);
}

export interface OperationContext extends langium.AstNode {
    readonly $container: OclDocument;
    readonly $type: 'OperationContext';
    conditions: Array<OperationCondition>;
    parameters: Array<Parameter>;
    path: QualifiedName;
    returnType?: TypeExpression;
}

export const OperationContext = 'OperationContext';

export function isOperationContext(item: unknown): item is OperationContext {
    return reflection.isInstance(item, OperationContext);
}

export interface Parameter extends langium.AstNode {
    readonly $container: DefExpression | OperationContext;
    readonly $type: 'Parameter';
    name: string;
    type: TypeExpression;
}

export const Parameter = 'Parameter';
// ...
export interface PropertyContext extends langium.AstNode {
    readonly $container: OclDocument;
    readonly $type: 'PropertyContext';
    constraints: Array<PropertyConstraint>;
    path: QualifiedName;
    propertyType: TypeExpression;
}

export const PropertyContext = 'PropertyContext';
// ...
export type OclLangiumAstType = {
    ArrowExpression: ArrowExpression
    BinaryExpression: BinaryExpression
    BodyCondition: BodyCondition
    BooleanLiteral: BooleanLiteral
    ClassifierConstraint: ClassifierConstraint
    ClassifierContext: ClassifierContext
    CoclContext: CoclContext
    CoclDerived: CoclDerived
    CoclReferenceFilter: CoclReferenceFilter
    CoclValidation: CoclValidation
    CollectionItem: CollectionItem
    CollectionLiteral: CollectionLiteral
    CollectionType: CollectionType
    ContextDeclaration: ContextDeclaration
    DefExpression: DefExpression
    DeriveConstraint: DeriveConstraint
    DotExpression: DotExpression
    Expression: Expression
    IfExpression: IfExpression
    InitConstraint: InitConstraint
    IntegerLiteral: IntegerLiteral
    InvalidLiteral: InvalidLiteral
    InvariantConstraint: InvariantConstraint
    IteratorBody: IteratorBody
    LetExpression: LetExpression
    NullLiteral: NullLiteral
    OclDocument: OclDocument
    OperationCondition: OperationCondition
    OperationContext: OperationContext
    Parameter: Parameter
    PostCondition: PostCondition
    PreCondition: PreCondition
    PropertyConstraint: PropertyConstraint
    PropertyContext: PropertyContext
    RealLiteral: RealLiteral
    SelfExpression: SelfExpression
    SimpleType: SimpleType
    StringLiteral: StringLiteral
    TupleLiteral: TupleLiteral
    TuplePart: TuplePart
    TupleType: TupleType
    TupleTypePart: TupleTypePart
    TypeExpLiteral: TypeExpLiteral
    TypeExpression: TypeExpression
    UnaryExpression: UnaryExpression
    VariableExpression: VariableExpression
}